% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistortedWeibull.R
\name{DistortedWeibull}
\alias{DistortedWeibull}
\alias{dDW}
\alias{pDW}
\alias{qDW}
\alias{rDW}
\title{Distorted Weibull distribution}
\usage{
rDW(n, parameters)

qDW(tau, parameters)

pDW(q, parameters)

dDW(x, parameters)
}
\arguments{
\item{n}{number of observations}

\item{parameters}{Weibull parameters}

\item{tau}{Belongs to parameters}

\item{x, q}{quantile}
}
\description{
Distorted Weibull distribution
}
\note{
TODO: Switch to named parameters
}

