% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistortedWeibull.R
\name{generateDWData}
\alias{generateDWData}
\title{Generates correlated, distorted weibull data}
\usage{
generateDWData(marginalParameters, copula, xpar)
}
\arguments{
\item{marginalParameters}{list of marginal parameters}

\item{copula}{copula object}

\item{xpar}{named list with `S` the number of sites
and `T` the number of days}
}
\description{
Generates correlated, distorted weibull data
}

