% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistortedWeibull.R
\name{setMarginalParameters}
\alias{setMarginalParameters}
\title{Get marginal parameters of the distorted Weibull distribution}
\usage{
setMarginalParameters(tau, epsilon, betas, kappas, gamma, xi)
}
\description{
Get marginal parameters of the distorted Weibull distribution
}

