% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistortedWeibull.R
\name{calculateCorrespondingWeibullShape}
\alias{calculateCorrespondingWeibullShape}
\title{Calculates the Weibull shape parameter}
\usage{
calculateCorrespondingWeibullShape(xi, tau)
}
\description{
that corresponds to a GPD shape parameter xi at quantile tau
}

