% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistortedWeibull.R
\name{setMarginalParameters}
\alias{setMarginalParameters}
\title{Get marginal parameters of the distorted Weibull distribution}
\usage{
setMarginalParameters(tau, epsilon, betas, kappas, gamma, xi)
}
\arguments{
\item{tau}{Quantile where GPD behaviors sets in}

\item{epsilon}{Length of transition interval}

\item{betas}{Weibull scale parameter}

\item{kappas}{Weibull shape parameter}

\item{gamma}{GPD dispersion coefficient}

\item{xi}{GPD shape parameter}
}
\description{
Get marginal parameters of the distorted Weibull distribution
}

